Add glossary to main notes document
The glossary (sans term definitions) has been appended to the document
however there are a few outstanding issues and choices to make regarding
this. Namely, should the referrences precede the glossary or vice versa?
Also, the Makefile now needs to be updated so that it correctly builds
both the glossary and the references. Running `make` at the moment will
not correctly build the full document.